+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.6] - 2024-12-12
+### Changed
+- Utilize Harbor for docker image management instead of docker registry
+- Add version_of_dev_env.sh to docker image
 
 ## [0.5.5] - 2024-12-09
 ### Added
@@ -73,3 +77,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.5.4]: https://github.com/potterwhite/DockerDevEnvTemplate/releases/tag/v0.5.4
 
 [0.5.5]: https://github.com/potterwhite/DockerDevEnvTemplate/releases/tag/v0.5.5
+
+[0.5.6]: https://github.com/potterwhite/DockerDevEnvTemplate/releases/tag/v0.5.6